+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Fix roomid / section not being specify consistently in room APIs
+
 ## 0.1.3 / 2012-12-25
 
 * Fix references from Message#text to Message#content in examples